Add missing sausage pies
Some of the expanded foods sausages were not matching the pie filling rule. So
update the rule to also include foods that have the pie attributes and can
smelt into another food that has nutritional properties.

Also add hints to describe which kind of sausage a pie was made from.