Support duplicate cooking recipes
Previously the cooking recipes were imported into the match rule's existing
slot conditions. A single match rule's slot conditions did not have the ability
to hold multiple imported recipes. So the code would return errors if multiple
recipes were found when it tried to import the recipe by name.

With this change, the imported recipes are held in a new condition inside the
match rule. If there are multiple recipes with the same name, this new
conditionunions the match results together.