Allow pies without a top crust
Pies that do not have a top crust have null for their index 5 content slot.
Change the pie matchers to allow them, in addition to allowing the top crust to
be present.

Add the NullForEmpty option to tell the matcher to fill in null for the
enumerated values if the min 0 is matched. This is necessary to prevent index
out of bound exceptions in the pie renderer when showing the enumerated pies in
the Gourmand tab.

Fixes https://github.com/bluelightning32/vs-gourmand/issues/8