Replace NotifyingMeal and NotifyingPie with Harmony patches
UpdateFoodAchievements.OnEntityReceiveSaturation is called whenever the player eats a food. However, it is not given the item stack of that was eaten, only the saturation stats of the food. Gourmand needs the item stack to see if it matches any of the food achievements. So Gourmand adds the NotifyEaten behavior to all food items. It intercepts the held interact stop event to tell the UpdateFoodAchievements behavior what food about to be eaten. Then OnEntityReceiveSaturation applies the food achievement, based on the item stack it was previously notified about.

However, the block class for meals and pies overrides OnHeldInteractionStop and bypasses all behaviors. Most likely, the VS team wrote the block class that way without thinking about the consequences.

Prior to this commit, Gourmand would subclass BlockMeal and BlockPie (it inherits the bad actions from BlockMeal) to override the Consume method to the UpdateFoodAchievements behavior before letting BlockMeal/BlockPie handle eating the food.

This caused a minor incompatibility with the Food Shelves mod. It looks at the block class to identify which blocks are pies. It needed to special case looking for both "pie" and "notifyingpie" when Gourmand was installed.

The bigger issue is that NotifyingPie and NotifyingMeal is causing an incompatibility with the Novelty mod. The Novelty mod forked those classes. The Novelty versions only call notify Novelty and skip Gourmand. So Gourmand stopped giving points for pies and meals.

So this commit uses Harmony to patch BlockMeal so that it properly calls behaviors. This commit also introduces new block behavior that intercepts the held interact stop event for meals/pies, along with several other events that proceed UpdateFoodAchievements.OnEntityReceiveSaturation.

The Harmony patch is designed such that it only installs itself once, even if Novelty ends up copying the code into its mod and both mods are installed. The patches also invokes all behaviors on the block. So if either mod applies the patch, then the behaviors for all mods will run.