Attempt the fix the rendering on the wrong thread crash
A crash was reported on ModDB where the Gourmand attribute was added through the GetHeldItemInfo call on the search cache thread. Adding the attribute triggered Gourmand to try to update the stat bars, but that is only allowed on the main thread.

The fix is to modifying the Gourmand attributes inside GetHeldItemInfo.